hhtc 2.1

+++ hhtc-mpp/src/main/java/com/jadyer/seed/mpp/web/service/MarketTransactionService.java
@@ -4,6 +4,7 @@ import com.jadyer.seed.comm.constant.CodeEnum;
 import com.jadyer.seed.comm.constant.CommonResult;
 import com.jadyer.seed.comm.constant.Constants;
 import com.jadyer.seed.comm.exception.HHTCException;
+import com.jadyer.seed.comm.util.DateUtil;
 import com.jadyer.seed.mpp.web.model.*;
 import com.jadyer.seed.mpp.web.repository.GoodsPublishOrderRepository;
 import com.jadyer.seed.mpp.web.repository.OrderInforRepository;
@@ -11,6 +12,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -114,20 +117,48 @@ public class MarketTransactionService {
     /**
      *TOKGO 用户预约下单
      * */
-    public String Reservation(String openid,String orderid,String CarNumber){
+    public String Reservation(String openid,String orderid,String CarNumber,int type){
         MppFansInfor fansInfor = fansService.getByOpenid(openid);
         if(fansInfor.getInfor_state().charAt(Constants.INFOR_STATE_CARNUMBE_BIT)!='1'){
             throw new HHTCException(CodeEnum.HHTC_INFOR_CARNUMBER_NO);
         }
-        //TODO  车牌可以多次预定吗
-        //进行订单预约
+        //获取订单
         GoodsPublishOrder goodsPublishOrder = goodsPublishOrderRepository.findByOrderID(orderid);
+        //订单时间冲突检查
+        if (type == 1&& OrderTimeCheck(goodsPublishOrder,openid)){
+            throw new HHTCException(CodeEnum.HHTC_ORDER_PORT_TIMEERROR);
+        }
+        //进行订单预约
         OrderInfor orderInfor = orderInforService.AddOrder(fansInfor,goodsPublishOrder,CarNumber);
         //市场消除订单
         goodsPublishOrderRepository.delete(goodsPublishOrder.getId());
         //TODO 微信模板消息通知车主
         //TODO 暂时没有返回
         return "";
+    }
+
+    /**
+     * TOKGO 订单时间冲突检查
+     * @return  true 含有时间冲突
+     * */
+    private boolean OrderTimeCheck(GoodsPublishOrder goodsPublishOrder,String openid){
+        try {
+            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm");
+            long starttime =  sdf.parse(goodsPublishOrder.getPublishFromTime()).getTime();
+            long endtime =  sdf.parse(goodsPublishOrder.getPublishEndTime()).getTime();
+            List<OrderInfor> orderInfors = orderInforService.GetOwnersOrder(openid);
+            for (OrderInfor orderInfor:orderInfors) {
+                long orderstarttime = sdf.parse(orderInfor.getTimeStart()).getTime();
+                long orderendtime = sdf.parse(orderInfor.getTimeEnd()).getTime();
+                if ((starttime>=orderstarttime&&starttime<orderendtime)
+                        || (starttime<=orderstarttime&&orderendtime<=endtime)
+                        || (endtime>=orderstarttime&&endtime<orderendtime))
+                    return true;
+            }
+        } catch (ParseException e) {
+            throw new HHTCException(CodeEnum.SYSTEM_ERROR);
+        }
+        return false;
     }
 
     /**
@@ -155,7 +186,7 @@ public class MarketTransactionService {
      * */
     public String StartOrder(String openid,String orderid){
         //获取订单
-        OrderInfor orderInfor = orderInforService.GetOrder(openid,orderid);
+        OrderInfor orderInfor = orderInforService.GetOrder(orderid);
         if (orderInfor==null)
             throw new HHTCException(CodeEnum.SYSTEM_PARAM_ERROR);
         if (!orderInforService.StartOrder(orderInfor)){
